time: use common implementation using gettimeofday()
Instead of providing yet another arm/__time.c file for ARM-EABI, this
patch merges existing time(2) implementations (which did all the same
thing but with different code) into one global source file.  Build of
them depends on the existence of the __NR_time symbol.

The x64_64 implementation of time.S was moved to __time.S and an empty
time.S file was added to build from the correct VPATH.