utime: replaced existing (broken) utime(2) wrappers with a common function
utime(2) is not available on every architecture. The alpha and ia64
provided already replacements, but these were wrong (they made utime(2)
an alias for utimes(2), which is wrong).

Patch adds a C implementation and a testprogram for this syscall. It
corrects also the type of the 2nd argument in the prototype which expects
a pointer to a *constant* buffer.

Signed-off-by: Enrico Scholz <enrico.scholz@informatik.tu-chemnitz.de>