fix detection of cross compiler toolchain for hppa and ppc
this was previously done by a rename of folders in the source to match
the target name of gcc, but this proved to be problematic for cases like
hppa, where both hppa1.1 and hppa2.0 use the same shortplatform in
dietlibc. so we instead check for these cases in diet.c and set
shortplatform accordingly.

for details see http://bugs.gentoo.org/221961